Ptr refinement, cmake mess

+++ src/PaymentStrategy.cpp
@@ -4,7 +4,7 @@ PaymentStrategy::PaymentStrategy() {
     // Constructor implementation
 }
 
-bool PaymentStrategy::processPayment(float amount) {
+bool PaymentStrategy::processPayment(double amount) {
     // Method to process payment
     return true; // Placeholder return value
 }
